requirements.txt
python-multipart
openpyxl
fastapi
uvicorn[standard]
pydantic
pyngrok
joblib
pandas
transformers
Pillow
python-multipart